[V0.1.9.0] The Economic API Update

+++ Valour/Server/Models/Economy/Currency.cs
@@ -41,7 +41,7 @@ public class Currency : ISharedCurrency
     /// <summary>
     /// The total amount of this currency that has been issued
     /// </summary>
-    public int Issued { get; set; }
+    public long Issued { get; set; }
 
     /// <summary>
     /// The number of decimal places this currency supports
